Update comentarios
se actualizan y agregan endpoints de comentarios

+++ src/main/java/com/conectaduoc/repository/PostCommentRepository.java
@@ -8,8 +8,8 @@ import com.conectaduoc.model.PostComment;
 
 public interface PostCommentRepository extends JpaRepository<PostComment, Long> {
     // Métodos personalizados para buscar comentarios por 
-    List<PostComment> findByIdPost(Integer idPost);
+    List<PostComment> findByIdPost(Long idPost);
     // Método para buscar comentarios por idUser
-    List<PostComment> findByIdUser(Integer idUser);
+    List<PostComment> findByIdUser(Long idUser);
 }
